feat(notifications): Ring for root agents, mark tabs in the sidebar
A subagent settles for one step of its parent's work, so a ring there
tells the user that their agent finished while the parent still runs.
The parent settles on its own when the whole piece of work is done, and
that is the settle worth interrupting somebody for. The dot that says
"this transcript moved" stays, and the workspace list now shows it too.

- Drop the turn-end sound for a subagent tab. The test runs BEFORE the
  60s cooldown, so a child cannot spend the ring the root's own settle
  needs a moment later. An agent whose lineage has not arrived rings but
  does not spend the cooldown, because it can still turn out to be a
  child
- Seed parentAgentId when openSubagentTab places the tab. Five consumers
  read that link, and each treats a tab without one as a root, so the
  window between the placement and the listAgents reply gave all five
  the wrong answer at once
- Extract isSubagentTab, and route the six call sites that spelled the
  parent-link test by hand through it
- Render the activity marker on a sidebar tab row, and roll it up to a
  collapsed workspace, repository or branch row, to a folded sidebar
  section and to the collapsed rail. Folding hides everything below it,
  which is the case the marker exists for. The roll-up resolves each tab
  through the live lookup, because the cached tree's fingerprint
  deliberately ignores the flag
- Move the dot into one NotificationDot component, so the tab strip and
  the sidebar cannot drift in look or markup, and give its test id a
  literal union so a typo is a compile error
- Restrict the turn-end refresh to the worker whose tree is on screen,
  and coalesce a burst: a parent and five subagents now cost two
  refreshes rather than six. Relevance decides, not lineage -- a
  background subagent's turn end is the only refresh its edits ever get.
  The trailing edge asks again, so a burst that went stale inside the
  window spends nothing and does not delay the turn end after it
- Extract the leading-edge throttle into ~/lib/throttle, beside the
  trailing debounce the project already owns
- Split the sidebar tree's pure model out of its component module. Five
  modules took their types from a renderer they sit beside
- Pin "a subagent turn end never settles its root" at each layer that
  could break it: the Codex child thread, the worker's activity
  derivation, and the client store the thinking indicator reads
